fix bug which calc repeat commission
