+++ trade_free/brain/base_brain.py
@@ -1,5 +1,8 @@
 import time
 
+import pandas as pd
+
+from risk_management.performance import create_sharpe_ratio, create_drawdowns
 from .abs_brain import AbsBrain
 from ..execution import SimulatedExecutionHandler
 from ..portfolio import SimplePortfolio
@@ -16,8 +19,9 @@ class BaseBrain(AbsBrain):
         super().__init__()
         self.bars = bars
         self.bars.register_event_queue(self.event_queue)
+        self.bars.init_data(start_date, end_date)
 
-        self.portfolio = SimplePortfolio(start_date, end_date, self.event_queue, self.bars, initial_capital)
+        self.portfolio = SimplePortfolio(start_date, self.event_queue, self.bars, initial_capital)
         self.broker = SimulatedExecutionHandler(self.event_queue)
 
     def add_Strategy(self, Strategy):
@@ -31,16 +35,15 @@ class BaseBrain(AbsBrain):
             else:
                 break
 
-            # Handle the event_queue
             while True:
                 try:
                     event = self.event_queue.pop()
                 except IndexError as e:
+                    self.portfolio.update_timeindex()  # 以最新的价格更新持仓信息
                     break
                 else:
                     if event:
                         if event.type == 'MARKET':  # 处理市场数据, 触发策略
-                            self.portfolio.update_timeindex()  # 以最新的价格更新持仓信息
                             self.strategy.calculate_signals(event)
 
                         elif event.type == 'SIGNAL':  # 策略执行, 触发订单
@@ -55,5 +58,34 @@ class BaseBrain(AbsBrain):
                         else:
                             raise ValueError("未知的event, event.type={0}".format(event.type))
 
-            # 10-Minute heartbeat
             time.sleep(0)
+
+        self.create_equity_curve_dataframe()
+
+    def create_equity_curve_dataframe(self):
+        """
+        创建一个数据体, 记录holding, 对后面数据分析有用
+        """
+        curve = pd.DataFrame(self.portfolio.all_holdings)
+        curve.set_index('datetime', inplace=True)
+        curve['returns'] = curve['total'].pct_change()
+        curve['equity_curve'] = (1.0 + curve['returns']).cumprod()
+        self.equity_curve = curve
+
+    def output_summary_stats(self):
+        """
+        展示收益统计信息
+        """
+        total_return = self.equity_curve['equity_curve'][-1]
+        returns = self.equity_curve['returns']
+        pnl = self.equity_curve['equity_curve']
+
+        sharpe_ratio = create_sharpe_ratio(returns)
+        max_dd, dd_duration = create_drawdowns(pnl)
+
+        stats = [("Total Return", "%0.2f%%" % ((total_return - 1.0) * 100.0)),
+                 ("Sharpe Ratio", "%0.2f" % sharpe_ratio),
+                 ("Max Drawdown", "%0.2f%%" % (max_dd * 100.0)),
+                 ("Drawdown Duration", "%d" % dd_duration)]
+        print(stats)
+        return stats
